+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to this project will be documented in this file.
 ### Added
 - add member list for attendance poll statistic
 - show unknown files on storage app on uploaded files page
+- show log files on developer page
 ### Fixed
 - fixed uploaded file page error on null storer files